[refactor] remove old experiments
